% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funs.R
\name{plot_one_country_year}
\alias{plot_one_country_year}
\title{Plot imprinting probabilities for a single country and year}
\usage{
plot_one_country_year(imprinting_df)
}
\arguments{
\item{imprinting_df}{A long data frame of imprinted probabilities output by \code{\link[=get_imprinting_probabilities]{get_imprinting_probabilities()}}. If the data frame contains more than one country and year, on the first will be plotted.}
}
\description{
Generate a stacked barplot, where each bar represents a birth cohort, and the colors within the bar show the probabilities that someone born in that cohort has a particular imprinting status.
}
\examples{
plot_one_country_year(imprinting_df = get_imprinting_probabilities(observation_years = 2022, countries = "Aruba"))

imprinting_df <- get_imprinting_probabilities(observation_years = 1997, countries = c("Algeria", "South Africa"))
plot_one_country_year(imprinting_df)
}
